Rust API: Fix all cargo clippy warnings
Applied automatic fixes and manual corrections for: unnecessary casts
left over from f64->f32 migration, doc comment overindentation,
map_or -> is_none_or/is_some_and, derivable Default impl for
FrameFilter, while_let_loop, iter().copied().collect() -> to_vec(),
and suppressed needless_range_loop where refactoring would reduce
readability.

+++ lib/rust/moteus-protocol/src/command.rs
@@ -573,7 +573,7 @@ impl OutputNearestCommand {
 
         writer.write_u8(multiplex::WRITE_FLOAT | 0x01);
         writer.write_varuint(Register::SetOutputNearest.address() as u32);
-        writer.write_f32(self.position as f32);
+        writer.write_f32(self.position);
     }
 }
 
@@ -595,7 +595,7 @@ impl OutputExactCommand {
 
         writer.write_u8(multiplex::WRITE_FLOAT | 0x01);
         writer.write_varuint(Register::SetOutputExact.address() as u32);
-        writer.write_f32(self.position as f32);
+        writer.write_f32(self.position);
     }
 }
